% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exquisite-package.R
\docType{package}
\name{exquisite-package}
\alias{exquisite}
\alias{exquisite-package}
\title{exquisite: Easily Squeeze URLs Using the TinyURL API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Exquisite allows users to squeeze or shorten long URLS into tiny ones using the TinyURL API. The package also allows users to update their exquisite new links by providing an alias,modifying tags,and setting an expiration date.
}
\author{
\strong{Maintainer}: JP Monteagudo \email{jpmonteagudo2014@gmail.com} (\href{https://orcid.org/0009-0003-6465-6658}{ORCID}) [copyright holder]

}
\keyword{internal}
