% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{recall}
\alias{recall}
\title{Retrieve Short and Long URL from TinyURL or iShortn}
\usage{
recall(domain = "tinyurl.com", alias = NULL)
}
\arguments{
\item{domain}{A character string specifying the domain to use for retrieving the shortened URL. It can either be \code{"tinyurl.com"} or \code{"ishortn.ink"}. Defaults to \code{"tinyurl.com"}.}

\item{alias}{A character string specifying the alias of the shortened URL. The alias must be at least 5 characters long.}
}
\value{
A list containing two elements:
\itemize{
\item \code{short_url}: The shortened URL.
\item \code{long_url}: The original long URL.
}
}
\description{
This function retrieves the \strong{shortened URL} and the \strong{original long URL} associated with an alias from either TinyURL or iShortn. The function works dynamically based on the provided domain (\code{tinyurl.com} or \code{ishortn.ink}).
}
\details{
The function sends a \strong{GET request} to either the TinyURL or iShortn API, depending on the specified domain. If the domain is \code{"tinyurl.com"}, it constructs a URL for the TinyURL API and retrieves both the short URL (\code{tiny_url}) and long URL (\code{url}). For \code{"ishortn.ink"}, it retrieves the short URL (\code{shortLink}) and long URL (\code{url}).

If the \code{alias} is not provided or is invalid, the function will stop with an error.
}
\examples{
\dontrun{
# Retrieve the short and long URL from TinyURL
tinyurl_info <- recall(domain = "tinyurl.com", alias = "myalias")
print(tinyurl_info)

# Retrieve the short and long URL from iShortn
ishortn_info <- recall(domain = "ishortn.ink", alias = "myalias")
print(ishortn_info)
}

}
